Create summitpost-into-google-earth.py
This script runs two loops which scrape the site and then builds a KML file with the results. The first loop scrapes a set of pages of search results containing all the mountain pages on the site. The second loop scrapes each mountain page and fills in details. The final section places the results as points in a KML file which can be viewed in Google Earth.